Changes to be committed: 	deleted:    .vscode/settings.json 	deleted:    build/LO21.exe 	modified:   include/linkedList.h

+++ include/linkedList.h
@@ -2,9 +2,9 @@
 #define LINKEDLIST_H
 
 /*
-*The linked list with void pointer so that the list
-*can take any type without rewriting the whole code each time
-*/
+ *The linked list with void pointer so that the list
+ *can take any type without rewriting the whole code each time
+ */
 typedef struct Node Node;
 struct Node
 {
@@ -18,16 +18,16 @@ struct LinkedList
 {
     Node *start;
     Node *end;
-    size_t listLength;
+    int listLength;
 };
 
-void LinkedList_Print(LinkedList*, void (*printFunc)(void*));
-int LinkedList_isEmpty(LinkedList*);
-void LinkedList_Free(LinkedList*);
-void LinkedList_Remove(LinkedList*, size_t);
-void LinkedList_Insert(LinkedList*,void*,size_t);
-void LinkedList_Append(LinkedList*,void*);
-void* LinkedList_Get(LinkedList*,size_t);
-LinkedList* createLinkedList();
+void LinkedList_Print(LinkedList *, void (*printFunc)(void *));
+int LinkedList_isEmpty(LinkedList *);
+void LinkedList_Free(LinkedList *);
+void LinkedList_Remove(LinkedList *, int);
+void LinkedList_Insert(LinkedList *, void *, int);
+void LinkedList_Append(LinkedList *, void *);
+void *LinkedList_Get(LinkedList *, int);
+LinkedList *createLinkedList();
 
 #endif
